Convenience functions

Makie also offers some convenience functions to make manipulating the Axis easier.

AbstractPlotting.xlims! — Function

xlims!(limits::Real...)
xlims!(limits::NTuple{2, Real})
xlims!(scene, limits::Real...)
xlims!(scene, limits::NTuple{2, Real})

Set the x-limits for the given Scene (defaults to current Scene).

AbstractPlotting.ylims! — Function

ylims!(limits::Real...)
ylims!(limits::NTuple{2, Real})
ylims!(scene, limits::Real...)
ylims!(scene, limits::NTuple{2, Real})

Set the y-limits for the given Scene (defaults to current Scene).

AbstractPlotting.zlims! — Function

zlims!(limits::Real...)
zlims!(limits::NTuple{2, Real})
zlims!(scene, limits::Real...)
zlims!(scene, limits::NTuple{2, Real})

Set the z-limits for the given Scene (defaults to current Scene).

AbstractPlotting.xlabel! — Function

xlabel!([scene,] xlabel)

Set the x-axis label for the given Scene. Defaults to using the current Scene.

AbstractPlotting.ylabel! — Function

ylabel!([scene,] ylabel)

Set the y-axis label for the given Scene. Defaults to using the current Scene.

AbstractPlotting.zlabel! — Function

zlabel!([scene,] zlabel)

Set the z-axis label for the given Scene. Defaults to using the current Scene.

Warning

The Scene must have an Axis3D. If not, then this function will error.

AbstractPlotting.xticklabels — Function

xticklabels(scene)

Returns the all the x-axis tick labels. See also ticklabels.

AbstractPlotting.yticklabels — Function

yticklabels(scene)

Returns the all the y-axis tick labels. See also ticklabels.

AbstractPlotting.zticklabels — Function

zticklabels(scene)

Returns the all the z-axis tick labels. See also ticklabels.

AbstractPlotting.xtickrange — Function

xtickrange(scene)

Returns the tick range along the x-axis. See also tickranges.

AbstractPlotting.ytickrange — Function

ytickrange(scene)

Returns the tick range along the y-axis. See also tickranges.

AbstractPlotting.ztickrange — Function

ztickrange(scene)

Returns the tick range along the z-axis. See also tickranges.

AbstractPlotting.xticks! — Function

xticks!([scene,]; xtickranges=xtickrange(scene), xticklabels=xticklabel(scene))

Set the tick labels and range along the x-axes. See also ticks!.

AbstractPlotting.yticks! — Function

yticks!([scene,]; ytickranges=ytickrange(scene), yticklabels=yticklabel(scene))

Set the tick labels and range along all the y-axis. See also ticks!.

AbstractPlotting.zticks! — Function

zticks!([scene,]; ztickranges=ztickrange(scene), zticklabels=zticklabel(scene))

Set the tick labels and range along all z-axis. See also ticks!.

AbstractPlotting.xtickrotation — Function

xtickrotation(scene)

Returns the rotation of tick labels along the x-axis. See also tickrotations

AbstractPlotting.ytickrotation — Function

ytickrotation(scene)

Returns the rotation of tick labels along the y-axis. See also tickrotations

AbstractPlotting.ztickrotation — Function

ztickrotation(scene)

Returns the rotation of tick labels along the z-axis. See also tickrotations

AbstractPlotting.xtickrotation! — Function

xtickrotation!([scene,] xangle)

Set the rotation of tick labels along the x-axis. See also tickrotations!.

AbstractPlotting.ytickrotation! — Function

ytickrotation!([scene,] yangle)

Set the rotation of tick labels along the y-axis. See also tickrotations!.

AbstractPlotting.ztickrotation! — Function

ztickrotation!([scene,] zangle)

Set the rotation of tick labels along the z-axis. See also tickrotations!.