Help functions

AbstractPlotting.help — Function

help(func[; extended = false])

Welcome to the main help function of Makie.jl / AbstractPlotting.jl.

For help on a specific function's arguments, type help_arguments(function_name).

For help on a specific function's attributes, type help_attributes(plot_Type).

Use the optional extended = true keyword argument to see more details.

Example usage:

help(scatter; extended = true)

scatter(positions)
scatter(x, y)
scatter(x, y, z)

Plots a marker for each element in (x, y, z), (x, y), or positions.

Theme

Available attributes and their defaults for Scatter{...} are:

  ambient           Float32[0.55, 0.55, 0.55]
  color             :black
  colormap          :viridis
  diffuse           Float32[0.4, 0.4, 0.4]
  distancefield     "nothing"
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  lightposition     :eyeposition
  linewidth         1
  marker            GeometryBasics.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  nan_color         RGBA{Float32}(0.0f0,0.0f0,0.0f0,0.0f0)
  overdraw          false
  rotations         Billboard()
  shininess         32.0f0
  specular          Float32[0.2, 0.2, 0.2]
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  transparency      false
  uv_offset_width   Float32[0.0, 0.0, 0.0, 0.0]
  visible           true

scatter has the following function signatures:

  (Vector, Vector)
  (Vector, Vector, Vector)
  (Matrix)

Please refer to convert_arguments to find the full list of accepted arguments

Available attributes and their defaults for Scatter{...} are:

  ambient           Float32[0.55, 0.55, 0.55]
  color             :black
  colormap          :viridis
  diffuse           Float32[0.4, 0.4, 0.4]
  distancefield     "nothing"
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  lightposition     :eyeposition
  linewidth         1
  marker            GeometryBasics.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  nan_color         RGBA{Float32}(0.0f0,0.0f0,0.0f0,0.0f0)
  overdraw          false
  rotations         Billboard()
  shininess         32.0f0
  specular          Float32[0.2, 0.2, 0.2]
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  transparency      false
  uv_offset_width   Float32[0.0, 0.0, 0.0, 0.0]
  visible           true

You can see usage examples of scatter by running:

example_database(scatter)

help_arguments

AbstractPlotting.help_arguments — Function

help_arguments([io], func)

Returns a list of signatures for function func.

Example usage:

help_arguments(stdout, scatter)

┌ Warning: argument help isn't currently implemented correctly
└ @ AbstractPlotting ~/.julia/packages/AbstractPlotting/jOgYQ/src/documentation/documentation.jl:63
`scatter` has the following function signatures:

```
  (Vector, Vector)
  (Vector, Vector, Vector)
  (Matrix)
```

help_attributes

AbstractPlotting.help_attributes — Function

help_attributes([io], Union{PlotType, PlotFunction}; extended = false)

Returns a list of attributes for the plot type Typ. The attributes returned extend those attributes found in the default_theme.

Use the optional keyword argument extended (default = false) to show in addition the default values of each attribute. usage:

>help_attributes(scatter)
	alpha
	color
	colormap
	colorrange
	distancefield
	glowcolor
	glowwidth
	linewidth
	marker
	marker_offset
	markersize
	overdraw
	rotations
	strokecolor
	strokewidth
	transform_marker
	transparency
	uv_offset_width
	visible

Example usage:

help_attributes(stdout, Scatter; extended = true)

Available attributes and their defaults for `Scatter{...}` are:

```
  ambient           Float32[0.55, 0.55, 0.55]
  color             :black
  colormap          :viridis
  diffuse           Float32[0.4, 0.4, 0.4]
  distancefield     "nothing"
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  lightposition     :eyeposition
  linewidth         1
  marker            GeometryBasics.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  nan_color         RGBA{Float32}(0.0f0,0.0f0,0.0f0,0.0f0)
  overdraw          false
  rotations         Billboard()
  shininess         32.0f0
  specular          Float32[0.2, 0.2, 0.2]
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  transparency      false
  uv_offset_width   Float32[0.0, 0.0, 0.0, 0.0]
  visible           true
```

Plot styling options

Use these functions to find out the styling options.

available_marker_symbols

AbstractPlotting.available_marker_symbols()

Marker Symbols:
    + => +
    diamond => ◆
    star4 => ✦
    star8 => ✷
    star6 => 🟋
    rtriangle => ▶
    x => x
    circle => ●
    pentagon => ⬟
    vline => ┃
    cross => ✚
    xcross => ❌
    rect => ■
    ltriangle => ◀
    dtriangle => ▼
    utriangle => ▲
    star5 => ★
    octagon => ⯄
    hline => ━
    hexagon => ⬢

available_gradients

AbstractPlotting.available_gradients()

Gradient Symbol/Strings:
    Accent
    Accent_3
    Accent_4
    Accent_5
    Accent_6
    Accent_7
    Accent_8
    Blues
    Blues_3
    Blues_4
    Blues_5
    Blues_6
    Blues_7
    Blues_8
    Blues_9
    BrBG
    BrBG_10
    BrBG_11
    BrBG_3
    BrBG_4
    BrBG_5
    BrBG_6
    BrBG_7
    BrBG_8
    BrBG_9
    BuGn
    BuGn_3
    BuGn_4
    BuGn_5
    BuGn_6
    BuGn_7
    BuGn_8
    BuGn_9
    BuPu
    BuPu_3
    BuPu_4
    BuPu_5
    BuPu_6
    BuPu_7
    BuPu_8
    BuPu_9
    CMRmap
    Dark2
    Dark2_3
    Dark2_4
    Dark2_5
    Dark2_6
    Dark2_7
    Dark2_8
    GnBu
    GnBu_3
    GnBu_4
    GnBu_5
    GnBu_6
    GnBu_7
    GnBu_8
    GnBu_9
    Greens
    Greens_3
    Greens_4
    Greens_5
    Greens_6
    Greens_7
    Greens_8
    Greens_9
    Greys
    Greys_3
    Greys_4
    Greys_5
    Greys_6
    Greys_7
    Greys_8
    Greys_9
    OrRd
    OrRd_3
    OrRd_4
    OrRd_5
    OrRd_6
    OrRd_7
    OrRd_8
    OrRd_9
    Oranges
    Oranges_3
    Oranges_4
    Oranges_5
    Oranges_6
    Oranges_7
    Oranges_8
    Oranges_9
    PRGn
    PRGn_10
    PRGn_11
    PRGn_3
    PRGn_4
    PRGn_5
    PRGn_6
    PRGn_7
    PRGn_8
    PRGn_9
    Paired
    Paired_10
    Paired_11
    Paired_12
    Paired_3
    Paired_4
    Paired_5
    Paired_6
    Paired_7
    Paired_8
    Paired_9
    Pastel1
    Pastel1_3
    Pastel1_4
    Pastel1_5
    Pastel1_6
    Pastel1_7
    Pastel1_8
    Pastel1_9
    Pastel2
    Pastel2_3
    Pastel2_4
    Pastel2_5
    Pastel2_6
    Pastel2_7
    Pastel2_8
    PiYG
    PiYG_10
    PiYG_11
    PiYG_3
    PiYG_4
    PiYG_5
    PiYG_6
    PiYG_7
    PiYG_8
    PiYG_9
    PuBu
    PuBuGn
    PuBuGn_3
    PuBuGn_4
    PuBuGn_5
    PuBuGn_6
    PuBuGn_7
    PuBuGn_8
    PuBuGn_9
    PuBu_3
    PuBu_4
    PuBu_5
    PuBu_6
    PuBu_7
    PuBu_8
    PuBu_9
    PuOr
    PuOr_10
    PuOr_11
    PuOr_3
    PuOr_4
    PuOr_5
    PuOr_6
    PuOr_7
    PuOr_8
    PuOr_9
    PuRd
    PuRd_3
    PuRd_4
    PuRd_5
    PuRd_6
    PuRd_7
    PuRd_8
    PuRd_9
    Purples
    Purples_3
    Purples_4
    Purples_5
    Purples_6
    Purples_7
    Purples_8
    Purples_9
    RdBu
    RdBu_10
    RdBu_11
    RdBu_3
    RdBu_4
    RdBu_5
    RdBu_6
    RdBu_7
    RdBu_8
    RdBu_9
    RdGy
    RdGy_10
    RdGy_11
    RdGy_3
    RdGy_4
    RdGy_5
    RdGy_6
    RdGy_7
    RdGy_8
    RdGy_9
    RdPu
    RdPu_3
    RdPu_4
    RdPu_5
    RdPu_6
    RdPu_7
    RdPu_8
    RdPu_9
    RdYlBu
    RdYlBu_10
    RdYlBu_11
    RdYlBu_3
    RdYlBu_4
    RdYlBu_5
    RdYlBu_6
    RdYlBu_7
    RdYlBu_8
    RdYlBu_9
    RdYlGn
    RdYlGn_10
    RdYlGn_11
    RdYlGn_3
    RdYlGn_4
    RdYlGn_5
    RdYlGn_6
    RdYlGn_7
    RdYlGn_8
    RdYlGn_9
    Reds
    Reds_3
    Reds_4
    Reds_5
    Reds_6
    Reds_7
    Reds_8
    Reds_9
    Set1
    Set1_3
    Set1_4
    Set1_5
    Set1_6
    Set1_7
    Set1_8
    Set1_9
    Set2
    Set2_3
    Set2_4
    Set2_5
    Set2_6
    Set2_7
    Set2_8
    Set3
    Set3_10
    Set3_11
    Set3_12
    Set3_3
    Set3_4
    Set3_5
    Set3_6
    Set3_7
    Set3_8
    Set3_9
    Spectral
    Spectral_10
    Spectral_11
    Spectral_3
    Spectral_4
    Spectral_5
    Spectral_6
    Spectral_7
    Spectral_8
    Spectral_9
    Wistia
    YlGn
    YlGnBu
    YlGnBu_3
    YlGnBu_4
    YlGnBu_5
    YlGnBu_6
    YlGnBu_7
    YlGnBu_8
    YlGnBu_9
    YlGn_3
    YlGn_4
    YlGn_5
    YlGn_6
    YlGn_7
    YlGn_8
    YlGn_9
    YlOrBr
    YlOrBr_3
    YlOrBr_4
    YlOrBr_5
    YlOrBr_6
    YlOrBr_7
    YlOrBr_8
    YlOrBr_9
    YlOrRd
    YlOrRd_3
    YlOrRd_4
    YlOrRd_5
    YlOrRd_6
    YlOrRd_7
    YlOrRd_8
    YlOrRd_9
    acton
    afmhot
    algae
    alpine
    amp
    aquamarine
    army
    atlantic
    auerbach
    aurora
    autumn
    autumn1
    avocado
    balance
    bamako
    batlow
    beach
    berlin
    bgy
    bgyw
    bilbao
    binary
    bjy
    bkr
    bky
    blackbody
    bluegreenyellow
    blues
    bluesreds
    bmw
    bone
    bone_1
    bosch_garden
    bosch_hell
    botticelli
    brass
    brg
    broc
    brocO
    browncyan
    buda
    bwr
    canaletto
    candy
    cezanne
    cherry
    cinferno
    cividis
    cmyk
    coffee
    colorwheel
    cool
    coolwarm
    copper
    cork
    corkO
    cubehelix
    curl
    cyclic1
    cyclic2
    cyclic3
    cyclic_grey_15_85_c0_n256
    cyclic_grey_15_85_c0_n256_s25
    cyclic_mrybm_35_75_c68_n256
    cyclic_mrybm_35_75_c68_n256_s25
    cyclic_mygbm_30_95_c78_n256
    cyclic_mygbm_30_95_c78_n256_s25
    cyclic_wrwbw_40_90_c42_n256
    cyclic_wrwbw_40_90_c42_n256_s25
    darkrainbow
    darkterrain
    darktest
    davos
    deep
    deepsea
    delta
    dense
    devon
    dimgreys
    diverging_bkr_55_10_c35_n256
    diverging_bky_60_10_c30_n256
    diverging_bwr_40_95_c42_n256
    diverging_bwr_55_98_c37_n256
    diverging_cwm_80_100_c22_n256
    diverging_gkr_60_10_c40_n256
    diverging_gwr_55_95_c38_n256
    diverging_gwv_55_95_c39_n256
    diverging_isoluminant_cjm_75_c23_n256
    diverging_isoluminant_cjm_75_c24_n256
    diverging_isoluminant_cjo_70_c25_n256
    diverging_linear_bjr_30_55_c53_n256
    diverging_linear_bjy_30_90_c45_n256
    diverging_rainbow_bgymr_45_85_c67_n256
    fall
    fire
    flag
    fruitpunch
    fuchsia
    gist_earth
    gist_gray
    gist_heat
    gist_ncar
    gist_rainbow
    gist_stern
    gist_yarg
    gnuplot
    gnuplot2
    gray
    gray1
    grayC
    grays
    grayyellow
    greenbrownterrain
    greenpink
    greens
    greys
    gwv
    haline
    hawaii
    heat
    hokusai
    holbein
    hot
    hsv
    ice
    imola
    inferno
    island
    isolum
    isoluminant_cgo_70_c39_n256
    isoluminant_cgo_80_c38_n256
    isoluminant_cm_70_c39_n256
    jet
    jet1
    julia_colorscheme
    kb
    kdc
    kg
    kgy
    klimt
    kr
    lajolla
    lake
    lapaz
    leonardo
    lightrainbow
    lighttemperaturemap
    lightterrain
    lighttest
    linear_bgy_10_95_c74_n256
    linear_bgyw_15_100_c67_n256
    linear_bgyw_15_100_c68_n256
    linear_blue_5_95_c73_n256
    linear_blue_95_50_c20_n256
    linear_bmw_5_95_c86_n256
    linear_bmw_5_95_c89_n256
    linear_bmy_10_95_c71_n256
    linear_bmy_10_95_c78_n256
    linear_gow_60_85_c27_n256
    linear_gow_65_90_c35_n256
    linear_green_5_95_c69_n256
    linear_grey_0_100_c0_n256
    linear_grey_10_95_c0_n256
    linear_kry_5_95_c72_n256
    linear_kry_5_98_c75_n256
    linear_kryw_0_100_c71_n256
    linear_kryw_5_100_c64_n256
    linear_kryw_5_100_c67_n256
    linear_ternary_blue_0_44_c57_n256
    linear_ternary_green_0_46_c42_n256
    linear_ternary_red_0_50_c52_n256
    lisbon
    magma
    matter
    mint
    munch
    neon
    nipy_spectral
    nuuk
    ocean
    oleron
    oslo
    oxy
    pastel
    pearl
    phase
    picasso
    pigeon
    pink
    plasma
    plum
    prism
    rainbow
    rainbow1
    rainbow2
    rainbow_bgyr_35_85_c72_n256
    rainbow_bgyr_35_85_c73_n256
    rainbow_bgyrm_35_85_c69_n256
    rainbow_bgyrm_35_85_c71_n256
    redblue
    redgreensplit
    reds
    redsblues
    rembrandt
    roma
    romaO
    rose
    rust
    sandyterrain
    seaborn_bright
    seaborn_bright6
    seaborn_colorblind
    seaborn_colorblind6
    seaborn_dark
    seaborn_dark6
    seaborn_deep
    seaborn_deep6
    seaborn_icefire_gradient
    seaborn_muted
    seaborn_muted6
    seaborn_pastel
    seaborn_pastel6
    seaborn_rocket_gradient
    seismic
    sienna
    solar
    southwest
    speed
    spring
    starrynight
    summer
    sun
    sunset
    tab10
    tab20
    tab20b
    tab20c
    temperaturemap
    tempo
    terrain
    thermal
    thermometer
    tofino
    tokyo
    turbid
    turbo
    turku
    twilight
    valentine
    vangogh
    vermeer
    vik
    vikO
    viridis
    watermelon
    winter

For other plot attributes and their usage, see the section Plot attributes.