Help functions

help

AbstractPlotting.help — Function

help(func[; extended = false])

Welcome to the main help function of Makie.jl / AbstractPlotting.jl.

For help on a specific function's arguments, type help_arguments(function_name).

For help on a specific function's attributes, type help_attributes(plot_Type).

Use the optional extended = true keyword argument to see more details.

Example usage:

help(scatter; extended = true)

scatter(positions)
scatter(x, y)
scatter(x, y, z)

Plots a marker for each element in (x, y, z), (x, y), or positions.

Theme

Available attributes and their defaults for Scatter{...} are:

  alpha             1.0
  color             :black
  colormap          :viridis
  colorrange        AbstractPlotting.Automatic()
  distancefield     "nothing"
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  linewidth         1
  marker            GeometryTypes.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  overdraw          false
  rotations         Billboard()
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  transparency      false
  uv_offset_width   Float32[0.0, 0.0, 0.0, 0.0]
  visible           true

scatter has the following function signatures:

  (Vector, Vector)
  (Vector, Vector, Vector)
  (Matrix)

Please refer to convert_arguments to find the full list of accepted arguments

Available attributes and their defaults for Scatter{...} are:

  alpha             1.0
  color             :black
  colormap          :viridis
  colorrange        AbstractPlotting.Automatic()
  distancefield     "nothing"
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  linewidth         1
  marker            GeometryTypes.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  overdraw          false
  rotations         Billboard()
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  transparency      false
  uv_offset_width   Float32[0.0, 0.0, 0.0, 0.0]
  visible           true

You can see usage examples of scatter by running:

example_database(scatter)

help_arguments

AbstractPlotting.help_arguments — Function

help_arguments([io], func)

Returns a list of signatures for function func.

Example usage:

help_arguments(stdout, scatter)

┌ Warning: argument help isn't currently implemented correctly
└ @ AbstractPlotting ~/.julia/packages/AbstractPlotting/ov5c9/src/documentation/documentation.jl:63
`scatter` has the following function signatures:

```
  (Vector, Vector)
  (Vector, Vector, Vector)
  (Matrix)
```

help_attributes

AbstractPlotting.help_attributes — Function

help_attributes([io], Union{PlotType, PlotFunction}; extended = false)

Returns a list of attributes for the plot type Typ. The attributes returned extend those attributes found in the default_theme.

Use the optional keyword argument extended (default = false) to show in addition the default values of each attribute. usage:

>help_attributes(scatter)
	alpha
	color
	colormap
	colorrange
	distancefield
	glowcolor
	glowwidth
	linewidth
	marker
	marker_offset
	markersize
	overdraw
	rotations
	strokecolor
	strokewidth
	transform_marker
	transparency
	uv_offset_width
	visible

Example usage:

help_attributes(stdout, Scatter; extended = true)

Available attributes and their defaults for `Scatter{...}` are:

```
  alpha             1.0
  color             :black
  colormap          :viridis
  colorrange        AbstractPlotting.Automatic()
  distancefield     "nothing"
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  linewidth         1
  marker            GeometryTypes.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  overdraw          false
  rotations         Billboard()
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  transparency      false
  uv_offset_width   Float32[0.0, 0.0, 0.0, 0.0]
  visible           true
```

Plot styling options

Use these functions to find out the styling options.

available_marker_symbols

AbstractPlotting.available_marker_symbols()

Marker Symbols:
    + => +
    diamond => ◆
    star4 => ✦
    star8 => ✷
    star6 => 🟋
    rtriangle => ▶
    x => x
    circle => ●
    pentagon => ⬟
    vline => ┃
    cross => ✚
    xcross => ❌
    rect => ■
    ltriangle => ◀
    dtriangle => ▼
    utriangle => ▲
    star5 => ★
    octagon => ⯄
    hline => ━
    hexagon => ⬢

available_gradients

AbstractPlotting.available_gradients()

Gradient Symbol/Strings:
    Accent
    Blues
    BrBG
    BuGn
    BuPu
    Dark2
    GnBu
    Greens
    Greys
    OrRd
    Oranges
    PRGn
    Paired
    Pastel1
    Pastel2
    PiYG
    PuBu
    PuBuGn
    PuOr
    PuRd
    Purples
    RdBu
    RdGy
    RdPu
    RdYlBu
    RdYlGn
    Reds
    Set1
    Set2
    Set3
    Spectral
    YlGn
    YlGnBu
    YlOrBr
    YlOrRd
    algae
    amp
    balance
    bgy
    bgyw
    bjy
    bkr
    bky
    blues
    bluesreds
    bmw
    cinferno
    cividis
    colorwheel
    coolwarm
    curl
    cyclic1
    cyclic2
    cyclic3
    darkrainbow
    darktest
    deep
    delta
    dense
    diff
    dimgreys
    fire
    grays
    grayscale
    greens
    greys
    gwv
    haline
    heat
    ice
    inferno
    isolum
    kb
    kdc
    kg
    kgy
    kr
    lightrainbow
    lighttest
    magma
    matter
    oxy
    phase
    plasma
    pu_or
    rain
    rainbow
    reds
    redsblues
    solar
    speed
    tarn
    tempo
    thermal
    topo
    turbid
    viridis

For other plot attributes and their usage, see the section Plot attributes.